updated to make GalyleoObject the centrall structure

+++ src/permissions.py
@@ -1,6 +1,6 @@
 from google.cloud import datastore
 import os
-import galyleo_storage 
+from galyleo_object import GalyleoObject
 
 # make sure the Google Cloud credentials are appropriately stored and accessed
 
@@ -29,17 +29,17 @@ class DatastoreManager:
     self.namespace = namespace
     self.client = datastore.Client(project=project, database=database, namespace=namespace)
 
-  def _get_object(self, url):
+  def _get_object(self, key):
     # utility to pull the object at URL url
-    key = self.client.key("URL", url)
+    key = self.client.key("key", key)
     object = self.client.get(key)
     return object
 
 
-  def _add_user(self, url, user):
+  def _add_user(self, key, user):
     # utility to add user user to the permission list of URL.
     with self.client.transaction(): # lock the datastore
-      object = self._get_object(url)
+      object = self._get_object(key)
       if object is None:
         return # should this throw an error?
       if user in set(object["users"]):
@@ -48,10 +48,10 @@ class DatastoreManager:
       object["users"].append(user)
       self.client.put(object)
 
-  def _delete_user(self, url, user):
+  def _delete_user(self, key, user):
     # utility to delete user user from the permission list of URL
     with self.client.transaction(): # lock the datastore
-      object = self._get_object(url)
+      object = self._get_object(key)
       if object is None:
         return # should this throw an error
       current_users = set(object["users"])
@@ -59,26 +59,28 @@ class DatastoreManager:
         current_users.remove(user)
         object["users"] = list(current_users)
         self.client.put(object)
+
       
-  def create_or_update_object(self, url, owner, share_list = None):
+  def create_or_update_object(self, galyleo_object, share_list = None):
     '''
     Create an object indexed by url url, owned by user owner, or, if there's an existing object,
     update its share list to share_list/
     Parameters:
-      url: URL for the object
-      owner: user who owns the   object
+      galyleo_object: the object to create
       share_list: list of users permittied to access the object
     Returns:
       None
     '''
     if share_list is None: share_list = []
     share_set = set(share_list)
+    owner = galyleo_object.owner
+    key = galyleo_object.object_key
     share_set.add(owner)
 
     with self.client.transaction():
-      object = self._get_object(url)
+      object = self._get_object(key)
       if object is None:
-        key = self.client.key("URL", url)
+        key = self.client.key("key", key)
         object = datastore.Entity(key)
       else:
         share_set = share_set.union(set(object["users"]))
@@ -93,44 +95,47 @@ class DatastoreManager:
       self.client.put(object)
 
       
-  def delete_object(self, url):
+  def delete_object(self, galyleo_object):
     '''
-    Delete the object with URL url.
+    Delete the galyleo obejct
     Parameters:
-      url: URL for the created object
+     galyleo_object: object to be deleted 
     '''
-    key = self.client.key("URL", url)
+    if galyleo_object is None: return
+    key = self.client.key("key", galyleo_object.object_key)
     if key is not None:
       self.client.delete(key)
     # Throw an error if the key is None?
 
 
-  def add_user_access(self, url, user_name): 
+  def add_user_access(self, galyleo_object, user_name): 
     '''
     Add user_name as an accessor of this object.  Note that if user_name is HUB or PUBLIC it won't delete 
     previous access, just in case HUB or PUBLIC is later deleted.
-    Does nothing if there is no stored object with this URL
+    If there is no stored object with this object's key, add the owner and user to the list
     Parameters:
-      url: URL for the object to add the user to
+      galyleo_object: the object to add the user to
       user_name: user to add
     Returns:
       Nothing
 
     '''
-    object = self._get_object(url)
+    key = galyleo_object.object_key
+    object = self._get_object(key)
     if object is None:
       # shouldn't happen....
-      return # should an error be thrown?  Should the entity be created?
+      self.create_or_update_object(galyleo_object, [user_name])
+      
     else:
-      self._add_user(url, user_name)
+      self._add_user(key, user_name)
 
 
-  def delete_user_access(self, url, user_name):
+  def delete_user_access(self, galyleo_object, user_name):
     '''
     Delete user_name as an accessor of this object.  Note that if user_name is the object owner, the item won't be deleted
     Does nothing if there is no stored object with this URL.
     Parameters:
-      url: URL for the object to remove  the user from
+      galyleo_object: the object to remove the user from
       user_name: user to remove
     Returns:
       Nothing
@@ -138,28 +143,47 @@ class DatastoreManager:
     # Note that we shouldn't be able to delete the access of the owner.
     # This requires keeping a separate field, owner, which we'll set in create
     # and this will check and do nothing on an attempt to delete owner.
-    object = self._get_object(url)
+    key = galyleo_object.object_key
+    object = self._get_object(key)
     if object is None:
       # shouldn't happen....
       return # should an error be thrown?  Should the entity be created?
     else:
-      self._delete_user(url, user_name)
+      self._delete_user(key, user_name)
 
-  def get_users(self, url):
+  def get_users(self, galyleo_object):
     '''
-    Return the set of users with access to this object.  Note that if HUB or PUBLIC has access to
-    this object, then that is the only item returned, even if others are stored
+    Return the set of users with access to this object.  
     Parameters:
-      url: URL for the object to get the user list for
+      galyleo_object: the GalyleoObject to get the user list for
     Returns:
       The set of users which have access to the object
     '''
-    object = self._get_object(url)
+    key = galyleo_object.object_key
+    object = self._get_object(key)
     if object is not None:
       return set(object['users'])
     else: return set()
-
-
+  
+  def set_users(self, galyleo_object, user_set):
+    '''
+    Sets the set of users with access to this object. 
+    Parameters:
+      galyleo_object: the GalyleoObject to get the user list for
+      user_set: the updated set of users
+    Returns:
+      Nothing
+    Side Effects:
+      Updates the user set.  If there is no object, creates it
+    '''
+    with self.client.transaction():
+        key = galyleo_object.object_key
+        object = self._get_object(key)
+    if object is None:
+      self.create_or_update_object(galyleo_object, list(user_set))
+    else:
+      object["users"] = list(user_set)
+      self.client.put(object)
 
 
   def clear(self):
@@ -167,7 +191,7 @@ class DatastoreManager:
     Delete ALL the objects in this namespace.  Primarily designed for test use
 
     '''
-    query = self.client.query(kind = "URL")
+    query = self.client.query(kind = "key")
     query.keys_only()
     keys = query.fetch()
     self.client.delete_multi(keys)
@@ -175,81 +199,108 @@ class DatastoreManager:
 #--------------------------------------------------------------------------
 # tests
 #---------------------------------------------------------------------------
-from pathlib import Path
-
-def setup_tests():
-  parent = str(Path(__file__).parent.parent)
-  key_path = os.path.join(parent, '.keys', 'galyleo-server-1105bf54b967.json')
-  os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = key_path
-  os.environ['GALYLEO_STORAGE_BUCKET'] = 'galyleo-server-test-bucket'
-  os.environ['GOOGLE_PROJECT'] = 'galyleo-server'
-  os.environ['GALYLEO_PERMISSIONS_DATABASE'] = 'galyleo-server'
-  os.environ['GALYLEO_PERMISSIONS_NAMESPACE'] = 'test'
-
 
 def test_create_object():
-  manager.create_or_update_object('t1', 'user1')
-  key = manager.client.key("URL", "t1")
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  key = manager.client.key("key", test_object.object_key)
   object = manager.client.get(key)
   assert(object is not None)
-  assert(object["users"] == ["user1"] )
-  manager.create_or_update_object('t1', 'user1')
-  object = manager.client.get(key)
-  assert(object is not None)
-  assert(object["users"] == ["user1"] )
-  manager.create_or_update_object('t1', 'user2')
-  object = manager.client.get(key)
-  assert(object is not None)
-  assert(set(object["users"]) == set(["user1", "user2"]))
+  assert(object["users"] == [test_object.owner] )
+  manager.client.close()
+  del manager
+ 
   
-  
+# test_create_object()
 
 def test_delete_object():
-  manager.create_or_update_object('t1', 'user1')
-  key = manager.client.key("URL", "t1")
-  manager.delete_object('t1')
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  key = manager.client.key("key", test_object.object_key)
+  object = manager.client.get(key)
+  assert(object is not None)
+  manager.delete_object(test_object)
   object = manager.client.get(key)
   assert object is None
 
+# test_delete_object()
 def test_add_user():
-  manager.create_or_update_object('t1', 'user1')
-  manager.add_user_access('t1', 'user1')
-  key = manager.client.key("URL", "t1")
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  key = manager.client.key("key", test_object.object_key)
   object = manager.client.get(key)
-  assert object["users"] == ['user1']
-  manager.add_user_access('t1', 'user2')
+  assert object["users"] == ['test']
+  manager.add_user_access(test_object, 'user2')
   object = manager.client.get(key)
-  assert set(object["users"]) == {'user1', 'user2'}
+  assert set(object["users"]) == {'test', 'user2'}
+  test_object_2 = GalyleoObject('table', 'test', 't2.table' )
+  manager.add_user_access(test_object_2, "user2")
+  key2 = manager.client.key("key", test_object_2.object_key)
+  object = manager.client.get(key2)
+  assert object is not None
+  assert set(object["users"]) == {"test", "user2"}
+# test_add_user()
 
 def test_delete_user():
-  manager.create_or_update_object('t1', 'user1')
-  manager.add_user_access('t1', 'user2')
-  key = manager.client.key("URL", "t1")
-  manager.delete_user_access('t1', 'user2')
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  key = manager.client.key("key", test_object.object_key)
   object = manager.client.get(key)
-  assert object["users"] == ['user1']
-  manager.delete_user_access('t1', 'user3')
+  assert object["users"] == ['test']
+  manager.add_user_access(test_object, 'user2')
+  manager.delete_user_access(test_object, 'user2')
+  assert object["users"] == ['test']
+  manager.delete_user_access(test_object, 'user3')
   object = manager.client.get(key)
-  assert object["users"] == ['user1']
+  assert object["users"] == ['test']
+
+
+# test_delete_user()
 
 def test_get_users():
-  manager.create_or_update_object('t1', 'user1')
-  assert manager.get_users("t1") == {"user1"}
-  manager.add_user_access('t1', 'user2')
-  assert manager.get_users("t1") == {"user1", "user2"}
-  
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  assert manager.get_users(test_object) == {"test"}
+  manager.add_user_access(test_object, 'user2')
+  assert manager.get_users(test_object) == {"test", "user2"}
 
+# test_get_users()
+
+def test_set_users():
+  manager = DatastoreManager('galyleo-server', 'galyleo-server', 'test')
+  manager.clear()
+  test_object = GalyleoObject('table', 'test', 't1.table' )
+  manager.create_or_update_object(test_object)
+  key = manager.client.key("key", test_object.object_key)
+  user_set = {"test", "user2", "user3"}
+  manager.set_users(test_object, user_set)
+  object = manager.client.get(key)
+  assert set(object["users"]) == user_set
+  test_object_2 = GalyleoObject('table', 'test', 't2.table' )
+  manager.set_users(test_object_2, user_set)
+  key2 = manager.client.key("key", test_object_2.object_key)
+  object = manager.client.get(key2)
+  assert object is not None
+  assert set(object["users"]) == user_set
+
+# test_set_users()
+  
+def run_tests():
+  test_create_object()
+  test_add_user()
+  test_delete_object()
+  test_delete_user()
+  test_get_users()
+  test_set_users()
  
-setup_tests()
-manager = DatastoreManager(os.environ['GOOGLE_PROJECT'], os.environ['GALYLEO_PERMISSIONS_DATABASE'], os.environ['GALYLEO_PERMISSIONS_NAMESPACE'])
-manager.clear()
-test_create_object()
-manager.clear()
-test_delete_object()
-manager.clear()
-test_add_user()
-manager.clear()
-test_delete_user()
-manager.clear()
-test_get_users()
-manager.clear()
+# run_tests()
